+++ example_scripts/run_bvas.py
@@ -1,56 +1,39 @@
-"""
-In this example script we show how to use BVASSelector to analyze data generated by data/covid_preprocessing.py.
-Note that we recommend running BVASSelector on GPU unless the number of alleles is moderate.
-"""
 import argparse
 
 import torch
 
 from bvas import BVASSelector
 
+if torch.cuda.is_available():
+    torch.set_default_tensor_type(torch.cuda.DoubleTensor)
+else:
+    torch.set_default_tensor_type(torch.DoubleTensor)
 
-def run(filename, S=50.0, nu_eff_multiplier=0.5, tau=100.0, T=500 * 1000, T_burnin=10 * 1000, seed=0):
-    """
-    :param str filename: Points to a `.pt` file generated by `data/covid_preprocessing.py`.
-    :param float S: The number of non-neutral alleles expected a priori.
-    :param float nu_eff_multiplier: Multiplier applied to the effective population size used in constructing
-        the `Y` and `Gamma` tensors that reside in the file specified by `filename`.
-    :param float tau: Controls the precision of the coefficients in the prior.
-    :param int T: Positive integer that controls the number of MCMC samples that are
-        generated (i.e. after burn-in/adaptation).
-    :param int T_burnin: Positive integer that controls the number of MCMC samples that are
-        generated during burn-in/adaptation.
-    :param int seed: Random number seed for reproducibility.
-    """
+
+def run(filename, S=50.0, nu_eff_multiplier=0.5, T=500 * 1000, tau=100.0, T_burnin=10 * 1000, seed=0):
     data = torch.load(filename)
     mutations = data['mutations']
 
     s = "Running inference on {} with {} regions and {} selection coefficients..."
     print(s.format(filename, data['num_regions'], len(mutations)))
 
-    if torch.cuda.is_available():
-        Y = data['Y'].cuda().double()
-        Gamma = data['Gamma'].cuda().double()
-        genotype_matrix = data['genotype'].cuda().double()
-    else:
-        Y = data['Y'].double()
-        Gamma = data['Gamma'].double()
-        genotype_matrix = data['genotype'].double()
-
-    selector = BVASSelector(Y,
-                            Gamma,
+    selector = BVASSelector(data['Y'].cuda().double(),
+                            data['Gamma'].cuda().double(),
                             mutations,
                             S=S,
                             nu_eff_multiplier=nu_eff_multiplier,
                             tau=tau,
-                            genotype_matrix=genotype_matrix,
+                            genotype_matrix=data['genotype'].cuda().double(),
                             variant_names=data['pango_idx'])
 
     selector.run(T=T, T_burnin=T_burnin, seed=seed)
     summary, growth_rates = selector.summary, selector.growth_rates
+
     print(summary.iloc[:20])
+    print('[pipsum] {:.2f}'.format(summary.PIP.sum().item()))
 
     tag = ".S_{}.nueff_{}.tau_{}.T_{}_{}.s{}".format(S, int(100 * nu_eff_multiplier), tau, T, T_burnin, seed)
+
     f = 'summary.' + filename[15:-3] + tag + '.csv'
     print("saving csv to {}".format(f))
     summary.to_csv(f)
@@ -63,9 +46,9 @@ def run(filename, S=50.0, nu_eff_multiplier=0.5, tau=100.0, T=500 * 1000, T_burn
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Run BVAS')
     parser.add_argument('--seed', type=int, default=0)
-    parser.add_argument('--T', type=int, default=500 * 1000)
+    parser.add_argument('--T', type=int, default=50 * 1000)
     parser.add_argument('--Tb', type=int, default=10 * 1000)
-    parser.add_argument('--nu-eff', type=float, default=0.5)
+    parser.add_argument('--nu-eff', type=float, default=1.0)
     parser.add_argument('--S', type=float, default=50.0)
     parser.add_argument('--tau', type=float, default=100.0)
     parser.add_argument('--filename', type=str)
